% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpt_indvs_mat.R
\name{add_cohort_chip_info}
\alias{add_cohort_chip_info}
\title{Adds Cohort and Chip information for the individuals in FinnGen}
\usage{
add_cohort_chip_info(pheno_data, ver)
}
\arguments{
\item{pheno_data}{A data.frame. The phenotype data. Needs at least column \code{ID}.}

\item{ver}{A string. The FinnGen version.}
}
\description{
Used for getting in the test individuals for each endpoint of the PRS data
in FinnGen. These are individuals not used in training the PRS.
}
\author{
Kira E. Detrois
}
