% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prs_files.R
\name{add_prs_col}
\alias{add_prs_col}
\title{Reads in PRS file and adds it as a column to the data.frame}
\usage{
add_prs_col(file_path, disease, col_map, prs_data)
}
\arguments{
\item{file_path}{A string (string). The path to the PRS file.}

\item{disease}{A string (string). The PRS endpoint description.}

\item{col_map}{A tibble. The mapping from endpoint names to PRS
endpoint description.}

\item{prs_data}{A tibble. The whole PRS data for all previous endpoints.}
}
\value{
A tibble. The \code{prs_data} with the added PRS results.
}
\description{
Reads in PRS file and adds it as a column to the data.frame
}
\author{
Kira E. Detrois
}
