% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cols.R
\name{check_cols}
\alias{check_cols}
\title{Checks the column names of a file against the expected columns}
\usage{
check_cols(expect_cols, cols, file_path)
}
\arguments{
\item{expect_cols}{A character (vector). The expected column names}

\item{cols}{A character (vector). The received column name.}

\item{file_path}{A character (vector). The file from which the column names come from.}
}
\value{
The expected column names.
}
\description{
Replaces the columns names and throws a warning for largely differing names.
}
\author{
Kira E. Detrois
}
