% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{custom_colors_brewer}
\alias{custom_colors_brewer}
\title{Custom color selection}
\usage{
custom_colors_brewer(N_colors)
}
\arguments{
\item{N_colors}{The number of colors}
}
\description{
Based on https://colorbrewer2.org/#type=diverging&scheme=RdBu&n=10
}
\author{
Kira E. Detrois
}
