% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Reads in all the data files}
\usage{
get_all_data(
  score_type,
  endpts = NULL,
  pheno_file_path = "",
  icd_file_path = "",
  atc_file_path = "",
  prs_dir_path = "",
  prs_endpts_map = NULL,
  phers_dir_path = "",
  phers_study_descr = NULL,
  zip_dir_path = ""
)
}
\arguments{
\item{score_type}{A string (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index, then the
\code{icd_file_path} should be provided.
\item \code{EI} for Elixhauser index, then the
\code{icd_file_path} should be provided.
\item \code{PRS} for polygenic risk score, then the
\code{prs_file_path} should be provided.
\item \code{PheRS} for PheRS results score, then the
\code{phers_file_path} should be provided.
\item \code{MED} for medication data, then the
\code{atc_file_path} should be provided.
\item \code{EDU} for educational level. The column for this needs to
be called \code{EDU} in the \code{pheno_data}.
\item \code{ZIP} for zip scores, then the
\code{zip_file_path} should be provided.
}}

\item{endpts}{A string (vector). The endpoints of interest.}

\item{pheno_file_path}{A string. Path to the phenotype data. Needs to be provided.
The file needs at least columns as described
in \link[IHRC:get_relevant_pheno_data_cols]{IHRC::get_relevant_pheno_data_cols}.}

\item{icd_file_path}{A string. Path to the ICD data. Needs at least the
columns as described in \link[IHRC:get_study_cci_data]{IHRC::get_study_cci_data}.}

\item{atc_file_path}{A string. Path to the ATC data. The file needs at least the
columns as described in \link[IHRC:get_study_med_data]{IHRC::get_study_med_data}.}

\item{prs_dir_path}{A string. Path to the directory containing the PRS files.
The file needs at least the columns as described in
\link[IHRC:get_prs_endpt_data]{IHRC::get_prs_endpt_data}.}

\item{prs_endpts_map}{A tibble. Mapping the endpoint names to the ones used
in the PRS files. If not provided uses standard
mapping from \link{get_endpts} to
\link{get_prs_endpt_descr}.}

\item{phers_dir_path}{A string. Path to the directory containing the PRS files.
The file needs at least the columns as described in
\link[IHRC:get_prs_endpt_data]{IHRC::get_prs_endpt_data}.}

\item{phers_study_descr}{A string. The study description in the PheRS files. If not
provided uses standard call to \link{get_phers_file_descr}.}

\item{zip_dir_path}{A string. Path to the ZIP score data. The file needs at least
the columns as described in \link[IHRC:get_zip_data]{IHRC::get_zip_data}.}
}
\value{
A list all the data. At least \code{pheno} will contain data.
All others will contain data if in the \code{score_type} list(\code{pheno},
\code{icd}, \code{atc}, \code{prs}, \code{phers}, \code{zip})
}
\description{
Reads in all the data files
}
\author{
Kira E. Detrois
}
