% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ana_setup_file.R
\name{get_endpts_indvs_mat}
\alias{get_endpts_indvs_mat}
\title{Reads in the endpoint individuals selection files}
\usage{
get_endpts_indvs_mat(setup, pheno_data)
}
\arguments{
\item{setup}{A list. The setup information from the setup file.}

\item{pheno_data}{A data.frame. The phenotype data. Needs at least column \code{ID}.}
}
\value{
A data.frame with the individuals that can be used for each endpoint.
Contains a column of individual IDs and a binary column for each endpoint.
}
\description{
Reads in the endpoint individuals selection files
}
\author{
Kira E. Detrois
}
