% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exposure_data.R
\name{get_exposure_data}
\alias{get_exposure_data}
\title{Gets individuals matching exposure window}
\usage{
get_exposure_data(long_data, exp_start = NULL, exp_end = NULL)
}
\arguments{
\item{long_data}{A data.frame with at least columns \code{ID}, and
\code{PRIMARY_ICD}.}

\item{exp_start}{A numeric. Start of the exposure period. Can be
used to restrict the timeframe on which the index
should be calculated. In this case the data.frame
needs column \code{Event_age}.
In case it is a vector has to have the length of
the number of rows in \code{long_data}.}

\item{exp_end}{A numeric. End of the exposure period.
Can be used to restrict the timeframe on which the index
should be calculated. In this case the data.frame
needs column \code{Event_age}.
In case it is a vector has to have the length of
the number of rows in \code{long_data}.}
}
\description{
Subsets the data.frame to only include the entries from individuals
inside the exposure period. By default gets all data from the
data.frame. The exposure has to be given as age. In this case the
data.frame also needs to have a column \code{Event_age}.
}
\details{
If either or both \code{exp_start} and \code{exp_end} are provided restricts
the entries to the exposure period for each individual.
}
\author{
Kira E. Detrois
}
