% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phers_file_descr.R
\name{get_phers_file_descr}
\alias{get_phers_file_descr}
\title{Gets the study setup file description for the PheRS folders}
\usage{
get_phers_file_descr(
  study_type = "backward",
  obs_end_date = as.Date("2019/01/01"),
  exp_len = 10,
  wash_len = 2,
  obs_len = 8
)
}
\arguments{
\item{study_type}{A character. Currently only "backward" possible.}

\item{obs_end_date}{A Date. The end of the observation period.}

\item{exp_len}{An integer. Length of the exposure period (in years).}

\item{wash_len}{An integer. Length of the washout period (in years).}

\item{obs_len}{An integer. Length of the observation period
(in years).}
}
\description{
Gets the study setup file description for the PheRS folders
}
\author{
Kira E. Detrois
}
