% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prs_endpt_descr.R
\name{get_prs_endpt_descr}
\alias{get_prs_endpt_descr}
\title{Standard analysis PRS file names for endpoints}
\usage{
get_prs_endpt_descr()
}
\description{
Standard analysis PRS file names for endpoints
}
\author{
Kira E. Detrois
}
