% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_types.R
\name{get_score_types}
\alias{get_score_types}
\title{Get list of score types for the analysis}
\usage{
get_score_types(score_type, create_score_combos = FALSE, bunch_phenos = FALSE)
}
\arguments{
\item{score_type}{A string (vector). The score types to be used for the analysis.}

\item{create_score_combos}{A boolean, whether or not to create all possible
score type combinations from the score_type vector.
see function \link{get_all_possible_score_type_combs}.}
}
\value{
A list. The score type combinations to be run in the analyses.
}
\description{
Uses \code{IUtils::get_all_possible_score_type_combs}, if \code{create_score_combos}
is set to TRUE and otherwise creates a list of length 2 with the
\code{score_type} and an empty vector.
}
\author{
Kira E. Detrois
}
