% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setup.R
\name{get_setup}
\alias{get_setup}
\title{Reads in the setup file in form of a list}
\usage{
get_setup(setup_file_path)
}
\arguments{
\item{setup_file_path}{A string. The path to the setup file.}
}
\description{
Reads in the setup file, using function \link{read_setup_file} and
then manipulates the list elements to the classes needed for the
\link[IHRC:run_surv_studies]{IHRC::run_surv_studies} funciton.
}
\author{
Kira E. Detrois
}
