% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpt_indvs_mat.R
\name{join_endpts_indvs_mats}
\alias{join_endpts_indvs_mats}
\title{Joins two matrices for selecting individuals for an endpoint}
\usage{
join_endpts_indvs_mats(
  crnt_endpts_indvs_mat,
  endpt,
  prev_endpts_indvs_mat = NULL,
  set_nas_true = FALSE
)
}
\arguments{
\item{crnt_endpts_indvs_mat}{A dataframe. Contains a column of individual
IDs and a binary column named i.e. "J10_ASTHMA_crnt"
if the current endpoint is "J10_ASTHMA".}

\item{endpt}{A string. The current endpoint of interest.}

\item{prev_endpts_indvs_mat}{A dataframe. Contains a column of individual
IDs and a binary column named by the endpoint.}

\item{set_nas_true}{A boolean. Defines whether missing individuals in either matrix
are set to TRUE or FALSE. Default: TRUE}
}
\description{
Joins two matrices for selecting individuals for an endpoint
}
\author{
Kira E. Detrois
}
