% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_atc_file.R
\name{read_atc_file}
\alias{read_atc_file}
\title{Reads in the ATC file in FinnGen}
\usage{
read_atc_file(atc_file_path)
}
\arguments{
\item{atc_file_path}{A string (string). The path to the file.}
}
\value{
A tibble the ATC data with columns \code{ID}, \code{ATC}, \code{WEIGHT}.
}
\description{
Combines the ATC codes with the weights defined in \code{inst/extdata/ATC_weights.txt}
}
\author{
Kira E. Detrois
}
