% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpt_indvs_mat.R
\name{read_finngen_endpts_indvs_mat}
\alias{read_finngen_endpts_indvs_mat}
\title{Reads in the test individuals for each endpoint of the PRS data in FinnGen}
\usage{
read_finngen_endpts_indvs_mat(pheno_data, endpts, ver)
}
\arguments{
\item{pheno_data}{A data.frame. The phenotype data. Needs at least column \code{ID}.}

\item{endpts}{A string (vector). The endpoints of interest.}

\item{ver}{A string. The FinnGen version. Used for function
\link{add_cohort_chip_info}}
}
\description{
These are individuals not used in training the PRS.
}
\author{
Kira E. Detrois
}
