% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pheno_file.R
\name{read_pheno_file}
\alias{read_pheno_file}
\title{Reads in the phenotype file according to INTERVENE standard}
\usage{
read_pheno_file(file_path = "", endpts = NULL)
}
\arguments{
\item{file_path}{A string. The path to the file.}

\item{endpts}{A string (vector). The endpoints of interest.}
}
\value{
A tibble with the longitudinal phenotype data
}
\description{
See \href{https://docs.google.com/document/d/1GbZszpPeyf-hyb0V_YDx828YbM7woh8OBJhvzkEwo2g/edit}{INTERVENE phenotype file definition v1}.
}
\author{
Kira E. Detrois
}
