% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phers_endpt_indvs_mat.R
\name{read_phers_endpt_indvs_mat}
\alias{read_phers_endpt_indvs_mat}
\title{Title}
\usage{
read_phers_endpt_indvs_mat(
  dir_path,
  indvs_ids = character(),
  set_nas_true = TRUE,
  endpts = NULL,
  study_descr = NULL,
  indv_status = "test",
  prev_endpt_indvs_mat = NULL
)
}
\arguments{
\item{dir_path}{A character. The path to the PheRS file directory.}

\item{indvs_ids}{A character (vector). All individuals IDs. This is useful
if the endpoint selection information is not known for all
all individuals.}

\item{set_nas_true}{A boolean. Defines whether missing individuals in either matrix
are set to TRUE or FALSE. Default: TRUE.}

\item{endpts}{A character (vector). The endpoints of interest.}

\item{prev_endpt_indvs_mat}{A dataframe. Contains a column of individual
IDs and a binary column named by the endpoint.}
}
\description{
Title
}
\author{
Kira E. Detrois
}
