% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpt_indvs_mat.R
\name{read_phers_endpts_indvs_mat}
\alias{read_phers_endpts_indvs_mat}
\title{Reads in the test individuals for each endpoint of the PheRS data}
\usage{
read_phers_endpts_indvs_mat(
  dir_path,
  indvs_ids = character(),
  set_nas_true = TRUE,
  endpts = NULL,
  study_descr = NULL,
  prev_endpts_indvs_mat = NULL
)
}
\arguments{
\item{dir_path}{A string. The path to the PheRS file directory.}

\item{indvs_ids}{A string (vector). All individuals IDs. This is useful
if the endpoint selection information is not known for all
all individuals.}

\item{set_nas_true}{A boolean. Defines whether missing individuals in either matrix
are set to TRUE or FALSE.}

\item{endpts}{A string (vector). The endpoints of interest.}

\item{study_descr}{A string. The study description in the PheRS files. If not
provided uses standard call to \link{get_phers_file_descr}.}

\item{prev_endpts_indvs_mat}{A dataframe. Contains a column of individual
IDs and a binary column named by the endpoint.}
}
\description{
Reads in all files in the form of a matrix, with columns \code{ID} and
boolean columns for each endpoint with the selected testing individuals
for each endpoint.
}
\author{
Kira E. Detrois
}
