% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prs_files.R
\name{read_prs_files}
\alias{read_prs_files}
\title{Reads the PRS files for the endpoints from a common directory}
\usage{
read_prs_files(dir_path, endpts = NULL, prs_endpt_descr = NULL)
}
\arguments{
\item{dir_path}{A character (string). The path to the directory.}

\item{endpts}{A character (vector). The endpoint names. If not provided
uses default endpoints from function \code{\link[=get_endpts]{get_endpts()}}.}

\item{prs_endpt_descr}{A character (vector). The PRS files endpoint
descriptions. If not provided, uses default endpoints
from function \code{\link[=prs_endpt_descr]{prs_endpt_descr()}}.

The files need to be named in a format \code{prsendptdescr_PRS_hm3.sscore}.
For example, for endpoint \code{J10_ASTHMA} the file would be named
\code{Asthma_PRS_hm3.sscore}.

For the PRS endpoint descriptions, see function \code{\link[=get_prs_endpt_descr]{get_prs_endpt_descr()}}.

It is possible to provide custom endpoints or PRS endpoint descriptions,
using the parameters \code{endpts} and/or \code{prs_endpt_descr}. However, make sure
that the two vectors are in the same order and length. For the default
endpoint vector see \code{\link[=get_endpts]{get_endpts()}}.

The PRS files should at least contain columns \code{IID} and \code{SCORE1_AVG}.}
}
\value{
A tibble with columns \code{ID} and a column for each selected endpoint
named \code{endpt_PRS} i.e. \code{J10_ASTHMA_PRS}.
}
\description{
Reads the PRS files for the endpoints from a common directory
}
\author{
Kira E. Detrois
}
