% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setup.R
\name{read_setup_file}
\alias{read_setup_file}
\title{Reads in the setup file in form of a list}
\usage{
read_setup_file(setup_file_path)
}
\arguments{
\item{setup_file_path}{A string. The path to the setup file.}
}
\description{
Reads in the setup file in form of a list
}
\author{
Kira E. Detrois
}
