% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip_files.R
\name{read_zip_files}
\alias{read_zip_files}
\title{Reads the PheRS files for the endpoints from a common directory}
\usage{
read_zip_files(dir_path, endpts = NULL)
}
\arguments{
\item{dir_path}{A string (string). The path to the directory.}

\item{endpts}{A string (vector). The endpoint names. If not provided
uses default endpoints from function \code{\link[=get_endpts]{get_endpts()}}.}
}
\value{
A tibble with columns \code{ID} and a column for each selected endpoint
named \code{endpt_PRS} i.e. \code{J10_ASTHMA_PRS}.
}
\description{
Reads the PheRS files for the endpoints from a common directory
}
\author{
Kira E. Detrois
}
