% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ana_setup_file.R
\name{run_ana_setup_file}
\alias{run_ana_setup_file}
\title{Runs standard setup survival model}
\usage{
run_ana_setup_file()
}
\arguments{
\item{score_type}{}
}
\description{
Depending on the score types selected sets the covariants. All models
use the covariates sex and year of birth. Analyses with `PRS´,
additionally use the PCs, and Batch.
}
\details{
The exposure period is always 10 years, if washout is selected the
period is 2 years and the observation period 8 years. Otherwise, the
observation period is 10 years.

If \code{PheRS} selected as a score, also gets the test individuals matrix
for each endpoints.
}
\author{
Kira E. Detrois
}
