% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setup.R
\name{split_vecs}
\alias{split_vecs}
\title{Splits all the strings in the setup file into vector of strings}
\usage{
split_vecs(setup)
}
\arguments{
\item{setup}{A list. The setup information from the setup file.}
}
\value{
The updated setup
}
\description{
If the strings in the list are in fact vectors splits them into a vector of strings.
}
\author{
Kira E. Detrois
}
