% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom theme for plottig}
\usage{
theme_custom(
  base_size = 18,
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Text size}

\item{base_line_size}{Line size}

\item{base_rect_size}{Rectangle size}
}
\description{
Based on the
}
\author{
Kira E. Detrois
}
