% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cols.R
\name{unify_strs}
\alias{unify_strs}
\title{Casts columns names to lower charachter and removes all special signs}
\usage{
unify_strs(str)
}
\arguments{
\item{str}{A character. The string to be cast}
}
\description{
Casts columns names to lower charachter and removes all special signs
}
\author{
Kira E. Detrois
}
