% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cols.R
\name{check_cols}
\alias{check_cols}
\title{Check Column Names}
\usage{
check_cols(expect_cols, cols, file_path)
}
\arguments{
\item{expect_cols}{A character (vector). The expected column names}

\item{cols}{A character (vector). The actual column name.}

\item{file_path}{A character (vector). The file from which the column names come from.}
}
\value{
A modified vector of column names with the expected column names.
}
\description{
Checks if certain columns with specific names are present in a data frame,
and if they are not present, print a warning message indicating that the column is missing.
}
\author{
Kira E. Detrois
}
