% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_types.R
\name{get_all_possible_score_type_combs}
\alias{get_all_possible_score_type_combs}
\title{Creates a list of all possible combinations of the selected scores}
\usage{
get_all_possible_score_type_combs(score_type)
}
\arguments{
\item{score_type}{A string (vector). The score types to be used for the analysis.}
}
\value{
A list. The score type combinations to be run in the analyses.
}
\description{
Creates a list of all possible combinations of the selected scores
}
\examples{
score_types <- get_all_possible_score_type_combs(c("PRS", "CCI", "PheRS"))
}
\author{
Kira E. Detrois
}
