% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phers_files.R
\name{read_phers_files}
\alias{read_phers_files}
\title{Reads the PheRS files for the endpoints from a common directory}
\usage{
read_phers_files(dir_path, study_descr = NULL, endpts = NULL)
}
\arguments{
\item{dir_path}{A character (string). The path to the directory.}

\item{endpts}{A character (vector). The endpoint names. If not provided
uses default endpoints from function \code{\link[=get_endpts]{get_endpts()}}.}

\item{prs_endpt_descr}{A character (vector). The PRS files endpoint
descriptions. If not provided, uses default endpoints
from function \code{\link[=prs_endpt_descr]{prs_endpt_descr()}}.}
}
\value{
A tibble with columns \code{ID} and a column for each selected endpoint
named \code{endpt_PRS} i.e. \code{J10_ASTHMA_PRS}.
}
\description{
Reads the PheRS files for the endpoints from a common directory
}
\author{
Kira E. Detrois
}
