% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prs_files.R
\name{read_prs_files}
\alias{read_prs_files}
\title{Reads the PRS files for the endpoints from a common directory}
\usage{
read_prs_files(
  dir_path,
  prs_endpts_map = NULL,
  prs_file_end = "_PRS_hm3.sscore",
  prs_score_col_name = "SCORE1_AVG",
  prs_id_col_name = "#IID"
)
}
\arguments{
\item{dir_path}{A string. The path to the PRS file directory.}

\item{prs_endpts_map}{A tibble. Mapping the endpoint names to the ones used
in the PRS files. If not provided uses standard
mapping from \link{get_endpts} to
\link{get_prs_endpt_descr}.}
}
\value{
A tibble with columns \code{ID} and a column for each selected endpoint
named \code{endpt_PRS} i.e. \code{J10_ASTHMA_PRS}.
}
\description{
The files need to be named in a format \code{Asthma_PRS_hm3.sscore},
for example, for endpoint \code{J10_ASTHMA}.
}
\details{
For the PRS endpoint descriptions, see function \link{get_prs_endpt_descr}.

The PRS files should at least contain columns \code{IID} and \code{SCORE1_AVG}.
}
\author{
Kira E. Detrois
}
