% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ana_setup_file.R
\name{run_ana_setup_file}
\alias{run_ana_setup_file}
\title{Runs analysis defined in a setup file}
\usage{
run_ana_setup_file(setup_file_path)
}
\arguments{
\item{setup_file_path}{A string. The path to the setup file.}
}
\description{
Reads in all the data, using function \link{get_all_data} and then runs function
\link[IHRC:run_surv_studies]{IHRC::run_surv_studies}.
}
\author{
Kira E. Detrois
}
